move nats config back to the messaging package
