+++ messaging/config.go
@@ -1,9 +1,11 @@
-package nconf
+package messaging
 
 import (
 	"fmt"
 	"strings"
 	"time"
+
+	"github.com/netlify/netlify-commons/nconf"
 
 	"github.com/pkg/errors"
 
@@ -28,10 +30,10 @@ type NatsAuth struct {
 }
 
 type NatsConfig struct {
-	TLS           *TLSConfig `mapstructure:"tls_conf"`
-	DiscoveryName string     `mapstructure:"discovery_name" split_words:"true"`
-	Servers       []string   `mapstructure:"servers"`
-	Auth          NatsAuth   `mapstructure:"auth"`
+	TLS           *nconf.TLSConfig `mapstructure:"tls_conf"`
+	DiscoveryName string           `mapstructure:"discovery_name" split_words:"true"`
+	Servers       []string         `mapstructure:"servers"`
+	Auth          NatsAuth         `mapstructure:"auth"`
 
 	// for streaming
 	ClusterID string `mapstructure:"cluster_id" split_words:"true"`
